Replace XML documentation with CSharpDocumentation
XML comments for the purpose of generating API documentation are being
temporarily replaced with CSharpDocumentation attributes. The reason is
that with XML documentation alone it is impossible to include type and
member declarations in the final API documentation. I'm using reflection
on the compiled UnityConsole assembly to retrieve member declarations
with accompanying CSharpDocumentation attributes and generate the
documentation from that. The plan is to return to XML comments once I'm
able to successfully weave XML documentation with reflected member
declarations.